+++ jmx/jmx-command-handling/src/test/java/uk/gov/justice/services/jmx/command/HandlerMethodValidatorTest.java
@@ -1,12 +1,14 @@
 package uk.gov.justice.services.jmx.command;
 
-import static java.util.Optional.empty;
-import static java.util.Optional.of;
+import static java.util.UUID.randomUUID;
 import static org.hamcrest.CoreMatchers.is;
+import static org.hamcrest.CoreMatchers.nullValue;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.junit.jupiter.api.Assertions.fail;
 
 import uk.gov.justice.services.jmx.api.InvalidHandlerMethodException;
+import uk.gov.justice.services.jmx.api.parameters.JmxCommandRuntimeParameters;
+import uk.gov.justice.services.jmx.api.parameters.JmxCommandRuntimeParameters.JmxCommandRuntimeParametersBuilder;
 
 import java.lang.reflect.Method;
 import java.util.UUID;
@@ -18,7 +20,7 @@ import org.mockito.junit.jupiter.MockitoExtension;
 
 @ExtendWith(MockitoExtension.class)
 public class HandlerMethodValidatorTest {
-                                                 
+
     @InjectMocks
     private HandlerMethodValidator handlerMethodValidator;
 
@@ -26,21 +28,27 @@ public class HandlerMethodValidatorTest {
     public void givenNoCommandRuntimeId_shouldAcceptValidHandlerMethod() throws Exception {
 
         final TestCommandHandler testCommandHandler = new TestCommandHandler();
+        final JmxCommandRuntimeParameters jmxCommandRuntimeParameters = new JmxCommandRuntimeParametersBuilder()
+                .build();
 
         final Method validHandlerMethod = getMethod("validHandlerMethodWithOutRuntimeId", testCommandHandler.getClass());
 
-        handlerMethodValidator.checkHandlerMethodIsValid(validHandlerMethod, testCommandHandler, empty());
+        handlerMethodValidator.checkHandlerMethodIsValid(validHandlerMethod, testCommandHandler, jmxCommandRuntimeParameters);
     }
 
     @Test
     public void givenNoCommandRuntimeId_shouldFailIfMethodNotPublic() throws Exception {
 
         final TestCommandHandler testCommandHandler = new TestCommandHandler();
+        final JmxCommandRuntimeParameters jmxCommandRuntimeParameters = new JmxCommandRuntimeParametersBuilder()
+                .build();
+
+        assertThat(jmxCommandRuntimeParameters.getCommandRuntimeId(), is(nullValue()));
 
         final Method invalidPrivateHandlerMethod = getMethod("invalidPrivateHandlerMethodWithOutRuntimeId", testCommandHandler.getClass());
 
         try {
-            handlerMethodValidator.checkHandlerMethodIsValid(invalidPrivateHandlerMethod, testCommandHandler, empty());
+            handlerMethodValidator.checkHandlerMethodIsValid(invalidPrivateHandlerMethod, testCommandHandler, jmxCommandRuntimeParameters);
             fail();
         } catch (final InvalidHandlerMethodException expected) {
             assertThat(expected.getMessage(), is("Handler method 'invalidPrivateHandlerMethodWithOutRuntimeId' on class 'uk.gov.justice.services.jmx.command.TestCommandHandler' is not public."));
@@ -51,11 +59,15 @@ public class HandlerMethodValidatorTest {
     public void givenNoCommandRuntimeId_shouldFailIfMethodHasNoParameters() throws Exception {
 
         final TestCommandHandler testCommandHandler = new TestCommandHandler();
+        final JmxCommandRuntimeParameters jmxCommandRuntimeParameters = new JmxCommandRuntimeParametersBuilder()
+                .build();
+
+        assertThat(jmxCommandRuntimeParameters.getCommandRuntimeId(), is(nullValue()));
 
         final Method invalidMissingParameterHandlerMethod = getMethod("invalidMissingParameterHandlerMethod", testCommandHandler.getClass());
 
         try {
-            handlerMethodValidator.checkHandlerMethodIsValid(invalidMissingParameterHandlerMethod, testCommandHandler, empty());
+            handlerMethodValidator.checkHandlerMethodIsValid(invalidMissingParameterHandlerMethod, testCommandHandler, jmxCommandRuntimeParameters);
             fail();
         } catch (final InvalidHandlerMethodException expected) {
             assertThat(expected.getMessage(), is("Invalid handler method 'invalidMissingParameterHandlerMethod' on class 'uk.gov.justice.services.jmx.command.TestCommandHandler'. Method should have 2 parameters. First of type 'uk.gov.justice.services.jmx.api.command.SystemCommand' and second of type 'java.util.UUID'."));
@@ -67,10 +79,15 @@ public class HandlerMethodValidatorTest {
 
         final TestCommandHandler testCommandHandler = new TestCommandHandler();
 
+        final JmxCommandRuntimeParameters jmxCommandRuntimeParameters = new JmxCommandRuntimeParametersBuilder()
+                .build();
+
+        assertThat(jmxCommandRuntimeParameters.getCommandRuntimeId(), is(nullValue()));
+
         final Method invalidTooManyParametersHandlerMethod = getMethod("invalidTooManyParametersHandlerMethod", testCommandHandler.getClass());
 
         try {
-            handlerMethodValidator.checkHandlerMethodIsValid(invalidTooManyParametersHandlerMethod, testCommandHandler, empty());
+            handlerMethodValidator.checkHandlerMethodIsValid(invalidTooManyParametersHandlerMethod, testCommandHandler, jmxCommandRuntimeParameters);
             fail();
         } catch (final InvalidHandlerMethodException expected) {
             assertThat(expected.getMessage(), is("Invalid handler method 'invalidTooManyParametersHandlerMethod' on class 'uk.gov.justice.services.jmx.command.TestCommandHandler'. Method should have 2 parameters. First of type 'uk.gov.justice.services.jmx.api.command.SystemCommand' and second of type 'java.util.UUID'."));
@@ -81,11 +98,15 @@ public class HandlerMethodValidatorTest {
     public void givenNoCommandRuntimeId_shouldFailIfMethodDoesNotHaveSystemCommandAsParameter() throws Exception {
 
         final TestCommandHandler testCommandHandler = new TestCommandHandler();
+        final JmxCommandRuntimeParameters jmxCommandRuntimeParameters = new JmxCommandRuntimeParametersBuilder()
+                .build();
+
+        assertThat(jmxCommandRuntimeParameters.getCommandRuntimeId(), is(nullValue()));
 
         final Method invalidNoSystemCommandHandlerMethod = getMethod("invalidNoSystemCommandHandlerMethod", testCommandHandler.getClass());
 
         try {
-            handlerMethodValidator.checkHandlerMethodIsValid(invalidNoSystemCommandHandlerMethod, testCommandHandler, empty());
+            handlerMethodValidator.checkHandlerMethodIsValid(invalidNoSystemCommandHandlerMethod, testCommandHandler, jmxCommandRuntimeParameters);
             fail();
         } catch (final InvalidHandlerMethodException expected) {
             assertThat(expected.getMessage(), is("Invalid handler method 'invalidNoSystemCommandHandlerMethod' on class 'uk.gov.justice.services.jmx.command.TestCommandHandler'. Method should have 2 parameters. First of type 'uk.gov.justice.services.jmx.api.command.SystemCommand' and second of type 'java.util.UUID'."));
@@ -96,11 +117,15 @@ public class HandlerMethodValidatorTest {
     public void givenNoCommandRuntimeId_shouldFailIfMethodDoesNotHaveCommandIdAsParameter() throws Exception {
 
         final TestCommandHandler testCommandHandler = new TestCommandHandler();
+        final JmxCommandRuntimeParameters jmxCommandRuntimeParameters = new JmxCommandRuntimeParametersBuilder()
+                .build();
+
+        assertThat(jmxCommandRuntimeParameters.getCommandRuntimeId(), is(nullValue()));
 
         final Method invalidNoCommandIdMethod = getMethod("invalidNoCommandIdMethod", testCommandHandler.getClass());
 
         try {
-            handlerMethodValidator.checkHandlerMethodIsValid(invalidNoCommandIdMethod, testCommandHandler, empty());
+            handlerMethodValidator.checkHandlerMethodIsValid(invalidNoCommandIdMethod, testCommandHandler, jmxCommandRuntimeParameters);
             fail();
         } catch (final InvalidHandlerMethodException expected) {
             assertThat(expected.getMessage(), is("Invalid handler method 'invalidNoCommandIdMethod' on class 'uk.gov.justice.services.jmx.command.TestCommandHandler'. Method should have 2 parameters. First of type 'uk.gov.justice.services.jmx.api.command.SystemCommand' and second of type 'java.util.UUID'."));
@@ -110,22 +135,30 @@ public class HandlerMethodValidatorTest {
     @Test
     public void givenCommandRuntimeId_shouldAcceptValidHandlerMethod() throws Exception {
 
+        final UUID commandRuntimeId = randomUUID();
         final TestCommandHandler testCommandHandler = new TestCommandHandler();
+        final JmxCommandRuntimeParameters jmxCommandRuntimeParameters = new JmxCommandRuntimeParametersBuilder()
+                .withCommandRuntimeId(commandRuntimeId)
+                .build();
 
         final Method validHandlerMethod = getMethod("validHandlerMethodWithRuntimeId", testCommandHandler.getClass());
 
-        handlerMethodValidator.checkHandlerMethodIsValid(validHandlerMethod, testCommandHandler, of(UUID.randomUUID()));
+        handlerMethodValidator.checkHandlerMethodIsValid(validHandlerMethod, testCommandHandler, jmxCommandRuntimeParameters);
     }
 
     @Test
     public void givenCommandRuntimeId_shouldFailIfMethodNotPublic() throws Exception {
 
+        final UUID commandRuntimeId = randomUUID();
         final TestCommandHandler testCommandHandler = new TestCommandHandler();
+        final JmxCommandRuntimeParameters jmxCommandRuntimeParameters = new JmxCommandRuntimeParametersBuilder()
+                .withCommandRuntimeId(commandRuntimeId)
+                .build();
 
         final Method invalidPrivateHandlerMethod = getMethod("invalidPrivateHandlerMethodWithRuntimeId", testCommandHandler.getClass());
 
         try {
-            handlerMethodValidator.checkHandlerMethodIsValid(invalidPrivateHandlerMethod, testCommandHandler, of(UUID.randomUUID()));
+            handlerMethodValidator.checkHandlerMethodIsValid(invalidPrivateHandlerMethod, testCommandHandler, jmxCommandRuntimeParameters);
             fail();
         } catch (final InvalidHandlerMethodException expected) {
             assertThat(expected.getMessage(), is("Handler method 'invalidPrivateHandlerMethodWithRuntimeId' on class 'uk.gov.justice.services.jmx.command.TestCommandHandler' is not public."));
@@ -135,12 +168,16 @@ public class HandlerMethodValidatorTest {
     @Test
     public void givenCommandRuntimeId_shouldFailIfMethodHasNoParameters() throws Exception {
 
+        final UUID commandRuntimeId = randomUUID();
         final TestCommandHandler testCommandHandler = new TestCommandHandler();
+        final JmxCommandRuntimeParameters jmxCommandRuntimeParameters = new JmxCommandRuntimeParametersBuilder()
+                .withCommandRuntimeId(commandRuntimeId)
+                .build();
 
         final Method invalidMissingParameterHandlerMethod = getMethod("invalidMissingParameterHandlerMethod", testCommandHandler.getClass());
 
         try {
-            handlerMethodValidator.checkHandlerMethodIsValid(invalidMissingParameterHandlerMethod, testCommandHandler, of(UUID.randomUUID()));
+            handlerMethodValidator.checkHandlerMethodIsValid(invalidMissingParameterHandlerMethod, testCommandHandler, jmxCommandRuntimeParameters);
             fail();
         } catch (final InvalidHandlerMethodException expected) {
             assertThat(expected.getMessage(), is("Invalid handler method 'invalidMissingParameterHandlerMethod' on class 'uk.gov.justice.services.jmx.command.TestCommandHandler'. Method should have 3 parameters. First of type 'uk.gov.justice.services.jmx.api.command.SystemCommand' and second of type 'java.util.UUID' and third of type 'java.util.UUID'."));
@@ -150,12 +187,16 @@ public class HandlerMethodValidatorTest {
     @Test
     public void givenCommandRuntimeId_shouldFailIfMethodHasTooManyParameters() throws Exception {
 
+        final UUID commandRuntimeId = randomUUID();
         final TestCommandHandler testCommandHandler = new TestCommandHandler();
+        final JmxCommandRuntimeParameters jmxCommandRuntimeParameters = new JmxCommandRuntimeParametersBuilder()
+                .withCommandRuntimeId(commandRuntimeId)
+                .build();
 
         final Method invalidTooManyParametersHandlerMethod = getMethod("invalidTooManyParametersHandlerMethod", testCommandHandler.getClass());
 
         try {
-            handlerMethodValidator.checkHandlerMethodIsValid(invalidTooManyParametersHandlerMethod, testCommandHandler, of(UUID.randomUUID()));
+            handlerMethodValidator.checkHandlerMethodIsValid(invalidTooManyParametersHandlerMethod, testCommandHandler, jmxCommandRuntimeParameters);
             fail();
         } catch (final InvalidHandlerMethodException expected) {
             assertThat(expected.getMessage(), is("Invalid handler method 'invalidTooManyParametersHandlerMethod' on class 'uk.gov.justice.services.jmx.command.TestCommandHandler'. Method should have 3 parameters. First of type 'uk.gov.justice.services.jmx.api.command.SystemCommand' and second of type 'java.util.UUID' and third of type 'java.util.UUID'."));
@@ -165,12 +206,16 @@ public class HandlerMethodValidatorTest {
     @Test
     public void givenCommandRuntimeId_shouldFailIfMethodDoesNotHaveSystemCommandAsParameter() throws Exception {
 
+        final UUID commandRuntimeId = randomUUID();
         final TestCommandHandler testCommandHandler = new TestCommandHandler();
+        final JmxCommandRuntimeParameters jmxCommandRuntimeParameters = new JmxCommandRuntimeParametersBuilder()
+                .withCommandRuntimeId(commandRuntimeId)
+                .build();
 
         final Method invalidNoSystemCommandHandlerMethod = getMethod("invalidNoSystemCommandHandlerMethod", testCommandHandler.getClass());
 
         try {
-            handlerMethodValidator.checkHandlerMethodIsValid(invalidNoSystemCommandHandlerMethod, testCommandHandler, of(UUID.randomUUID()));
+            handlerMethodValidator.checkHandlerMethodIsValid(invalidNoSystemCommandHandlerMethod, testCommandHandler, jmxCommandRuntimeParameters);
             fail();
         } catch (final InvalidHandlerMethodException expected) {
             assertThat(expected.getMessage(), is("Invalid handler method 'invalidNoSystemCommandHandlerMethod' on class 'uk.gov.justice.services.jmx.command.TestCommandHandler'. Method should have 3 parameters. First of type 'uk.gov.justice.services.jmx.api.command.SystemCommand' and second of type 'java.util.UUID' and third of type 'java.util.UUID'."));
@@ -180,12 +225,16 @@ public class HandlerMethodValidatorTest {
     @Test
     public void givenCommandRuntimeId_shouldFailIfMethodDoesNotHaveCommandIdAsParameter() throws Exception {
 
+        final UUID commandRuntimeId = randomUUID();
         final TestCommandHandler testCommandHandler = new TestCommandHandler();
+        final JmxCommandRuntimeParameters jmxCommandRuntimeParameters = new JmxCommandRuntimeParametersBuilder()
+                .withCommandRuntimeId(commandRuntimeId)
+                .build();
 
         final Method invalidNoCommandIdMethod = getMethod("invalidNoCommandRuntimeIdMethod", testCommandHandler.getClass());
 
         try {
-            handlerMethodValidator.checkHandlerMethodIsValid(invalidNoCommandIdMethod, testCommandHandler, of(UUID.randomUUID()));
+            handlerMethodValidator.checkHandlerMethodIsValid(invalidNoCommandIdMethod, testCommandHandler, jmxCommandRuntimeParameters);
             fail();
         } catch (final InvalidHandlerMethodException expected) {
             assertThat(expected.getMessage(), is("Invalid handler method 'invalidNoCommandRuntimeIdMethod' on class 'uk.gov.justice.services.jmx.command.TestCommandHandler'. Method should have 3 parameters. First of type 'uk.gov.justice.services.jmx.api.command.SystemCommand' and second of type 'java.util.UUID' and third of type 'java.util.UUID'."));
@@ -194,8 +243,8 @@ public class HandlerMethodValidatorTest {
 
     private Method getMethod(final String methodName, final Class<?> handlerClass) {
 
-        for(final Method method: handlerClass.getDeclaredMethods()) {
-            if(method.getName().equals(methodName)) {
+        for (final Method method : handlerClass.getDeclaredMethods()) {
+            if (method.getName().equals(methodName)) {
                 return method;
             }
         }